+++ src/test/groovy/edu/ie3/simona/api/data/container/ExtInputContainerTest.groovy
@@ -3,6 +3,7 @@ package edu.ie3.simona.api.data.container
 import edu.ie3.datamodel.models.value.PValue
 import edu.ie3.simona.api.data.model.em.EmSetPoint
 import edu.ie3.simona.api.data.model.em.FlexOptionRequest
+import edu.ie3.simona.api.data.model.em.MultiFlexOptions
 import edu.ie3.simona.api.data.model.em.PowerLimitFlexOptions
 import spock.lang.Shared
 import spock.lang.Specification
@@ -12,8 +13,34 @@ import static edu.ie3.util.quantities.PowerSystemUnits.KILOWATT
 
 class ExtInputContainerTest extends Specification {
 
-    @Shared
-    private UUID sender = UUID.randomUUID()
+    def "An ExtInputContainer should return the tick correctly"() {
+        expect:
+        container.tick == expectedTick
+        container.maybeNextTick == expectedMaybeNextTick
+
+        where:
+        container                       | expectedTick | expectedMaybeNextTick
+        new ExtInputContainer(0L)       | 0L           | Optional.empty()
+        new ExtInputContainer(0L, 900L) | 0L           | Optional.of(900L)
+    }
+
+    def "An ExtInputContainer should check if it is empty correctly"() {
+        expect:
+        def container1 = new ExtInputContainer(0L)
+        container1.empty
+        container1.addPrimaryValue(UUID.randomUUID(), null)
+        !container1.empty
+
+        def container2 = new ExtInputContainer(0L)
+        container2.empty
+        container2.addFlexComMessage(null)
+        !container2.empty
+
+        def container3 = new ExtInputContainer(0L)
+        container3.empty
+        container3.addRequest(UUID.randomUUID())
+        !container3.empty
+    }
 
     def "An ExtInputContainer should add primary data correctly"() {
         given:
@@ -52,6 +79,21 @@ class ExtInputContainerTest extends Specification {
 
         when:
         container.addFlexOptions(receiver, [flexOptions])
+
+        then:
+        container.flexOptions == [(receiver): [flexOptions]]
+    }
+
+    def "An ExtInputContainer should add multi flex option data correctly"() {
+        given:
+        UUID receiver = UUID.randomUUID()
+        UUID sender = UUID.randomUUID()
+        def flexOptions = new PowerLimitFlexOptions(receiver, sender, Quantities.getQuantity(0d, KILOWATT), Quantities.getQuantity(2d, KILOWATT), Quantities.getQuantity(5d, KILOWATT))
+
+        def container = new ExtInputContainer(0L)
+
+        when:
+        container.addFlexOptions(new MultiFlexOptions(receiver, [(sender): flexOptions]))
 
         then:
         container.flexOptions == [(receiver): [flexOptions]]